requirements.txt
streamlit
openai
google-generativeai
google-api-python-client
gspread
oauth2client
pandas
matplotlib
reportlab
yt-dlp
google-auth-oauthlib
PyMuPDF 